Refactor TimeRange to App Level on Frontend
